refactor(cli): move out to its own packages
