+++ openllm-python/pyproject.toml
@@ -75,14 +75,14 @@ license = "Apache-2.0"
 name = "openllm"
 requires-python = ">=3.8"
 [project.scripts]
-openllm = "openllm.cli.entrypoint:cli"
-openllm-build-base-container = "openllm.cli.extension.build_base_container:cli"
-openllm-dive-bentos = "openllm.cli.extension.dive_bentos:cli"
-openllm-get-containerfile = "openllm.cli.extension.get_containerfile:cli"
-openllm-get-prompt = "openllm.cli.extension.get_prompt:cli"
-openllm-list-bentos = "openllm.cli.extension.list_bentos:cli"
-openllm-list-models = "openllm.cli.extension.list_models:cli"
-openllm-playground = "openllm.cli.extension.playground:cli"
+openllm = "openllm_cli.entrypoint:cli"
+openllm-build-base-container = "openllm_cli.extension.build_base_container:cli"
+openllm-dive-bentos = "openllm_cli.extension.dive_bentos:cli"
+openllm-get-containerfile = "openllm_cli.extension.get_containerfile:cli"
+openllm-get-prompt = "openllm_cli.extension.get_prompt:cli"
+openllm-list-bentos = "openllm_cli.extension.list_bentos:cli"
+openllm-list-models = "openllm_cli.extension.list_models:cli"
+openllm-playground = "openllm_cli.extension.playground:cli"
 
 [project.urls]
 Blog = "https://modelserving.com"
@@ -136,7 +136,7 @@ root = ".."
 [tool.hatch.metadata]
 allow-direct-references = true
 [tool.hatch.build.targets.wheel]
-only-include = ["src/openllm"]
+only-include = ["src/openllm", "src/openllm_cli"]
 sources = ["src"]
 [tool.hatch.build.targets.sdist]
 exclude = [
