+++ openllm-python/src/openllm_cli/extension/dive_bentos.py
@@ -12,9 +12,9 @@ from simple_di import inject
 import bentoml
 
 from bentoml._internal.configuration.containers import BentoMLContainer
-from openllm.cli import termui
-from openllm.cli._factory import bento_complete_envvar
-from openllm.cli._factory import machine_option
+from openllm_cli import termui
+from openllm_cli._factory import bento_complete_envvar
+from openllm_cli._factory import machine_option
 
 
 if t.TYPE_CHECKING:
