+++ openllm-python/src/openllm_cli/extension/get_containerfile.py
@@ -12,8 +12,8 @@ from bentoml._internal.bento.bento import BentoInfo
 from bentoml._internal.bento.build_config import DockerOptions
 from bentoml._internal.configuration.containers import BentoMLContainer
 from bentoml._internal.container.generate import generate_containerfile
-from openllm.cli import termui
-from openllm.cli._factory import bento_complete_envvar
+from openllm_cli import termui
+from openllm_cli._factory import bento_complete_envvar
 from openllm_core.utils import converter
